feat route protection by role

+++ src/app/core/services/auth.service.ts
@@ -1,30 +1,41 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject, PLATFORM_ID } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
 
-  constructor(private http: HttpClient) { }
+  private platformId = inject(PLATFORM_ID);
+
+  constructor(private http: HttpClient) {}
 
   login(data: any) {
     return this.http.post<any>('http://localhost:5224/api/Usuario/login', data);
   }
 
   guardarUsuario(user: any) {
-    localStorage.setItem('user', JSON.stringify(user));
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.setItem('user', JSON.stringify(user));
+    }
   }
 
   obtenerUsuario() {
-    return JSON.parse(localStorage.getItem('user') || 'null');
+    if (isPlatformBrowser(this.platformId)) {
+      const user = localStorage.getItem('user');
+      return user ? JSON.parse(user) : null;
+    }
+    return null;
   }
 
   isLogged() {
-    return !!this.obtenerUsuario();
+    return this.obtenerUsuario() !== null;
   }
 
   logout() {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
   }
 }
